TST: implement test_first (#32274)
