+++ pandas/tests/series/test_timeseries.py
@@ -167,60 +167,6 @@ class TestTimeSeries:
         expected = rng.get_indexer(ts_slice.index)
         tm.assert_numpy_array_equal(result, expected)
 
-    def test_first_subset(self):
-        ts = _simple_ts("1/1/2000", "1/1/2010", freq="12h")
-        result = ts.first("10d")
-        assert len(result) == 20
-
-        ts = _simple_ts("1/1/2000", "1/1/2010")
-        result = ts.first("10d")
-        assert len(result) == 10
-
-        result = ts.first("3M")
-        expected = ts[:"3/31/2000"]
-        tm.assert_series_equal(result, expected)
-
-        result = ts.first("21D")
-        expected = ts[:21]
-        tm.assert_series_equal(result, expected)
-
-        result = ts[:0].first("3M")
-        tm.assert_series_equal(result, ts[:0])
-
-    def test_first_raises(self):
-        # GH20725
-        ser = pd.Series("a b c".split())
-        msg = "'first' only supports a DatetimeIndex index"
-        with pytest.raises(TypeError, match=msg):
-            ser.first("1D")
-
-    def test_last_subset(self):
-        ts = _simple_ts("1/1/2000", "1/1/2010", freq="12h")
-        result = ts.last("10d")
-        assert len(result) == 20
-
-        ts = _simple_ts("1/1/2000", "1/1/2010")
-        result = ts.last("10d")
-        assert len(result) == 10
-
-        result = ts.last("21D")
-        expected = ts["12/12/2009":]
-        tm.assert_series_equal(result, expected)
-
-        result = ts.last("21D")
-        expected = ts[-21:]
-        tm.assert_series_equal(result, expected)
-
-        result = ts[:0].last("3M")
-        tm.assert_series_equal(result, ts[:0])
-
-    def test_last_raises(self):
-        # GH20725
-        ser = pd.Series("a b c".split())
-        msg = "'last' only supports a DatetimeIndex index"
-        with pytest.raises(TypeError, match=msg):
-            ser.last("1D")
-
     def test_format_pre_1900_dates(self):
         rng = date_range("1/1/1850", "1/1/1950", freq="A-DEC")
         rng.format()
